DataScienceEcosystem
In this notebook, Data Science Tools and Ecosystem are summarized.
We will show how create markdown cells, places titles, make list and create tables. We will also show how to enter an arithmetic expression, how to place codes to multiply and add numbers, finally we will convert minutes to hours and upload the notebook to GitHub.