Final Python Project
This file is a final Python project that was developed through the Coursera platform. The project is based on extracting financial information from the companies Tesla and GameStop, the closing prices of the shares are extracted with the *yfinance* library and the quarterly revenue of the same companies is extracted by webscrapping with the *BeautifulSoup* library, finally the information is displayed in a DashBoard through a proprietary function for graphing that allows you to interact with the date range.